Omit resources that have no methods
Currently resources that have no methods appear in the output
as hot links. When you click on an empty resource, it expands
to reveal that there is no content.

Note: It is possible that one of these empty resources
might have some documentation or a displayName. It would be
better to test for methods or a displayName or a description,
however I couldn't figure out an elegant way to have
handlebars do that.