IE11 problem opening modal a second time
Hi,

I have seen there is a problem when you try to open the same method modal view for a second time in a row. Apparently the 'hashchange' event is not received by the bind method. By this reason I have decided to set the window.location.href to the root url when the user hides the modal dialog. I don't know if the solution is the best, but it has worked for me.

Besides, I think the 'var anchor_id' is not needed in the 'hashchange' bind.

Regards